tjuln in nasprotnik sta dedovana

+++ trki.cpp
@@ -83,10 +83,8 @@ void GameManager::trkiMiTjulni() {
 	for (unsigned int i = 0; i < tjulni.size(); ++i) {
 		if (isPlayerCollidingAt(tjulni.at(i)->getX(), tjulni.at(i)->getY(), tjulni.at(i)->getW(), tjulni.at(i)->getH())) {
 			printf("Hvala lepa, resil si me!!!\n");
-			/*
-			*/
 			for (unsigned int popravi_nasprotnike = 0; popravi_nasprotnike < enemy.size(); ++popravi_nasprotnike) {
-				if (enemy.at(popravi_nasprotnike)->nosim > i && enemy.at(popravi_nasprotnike)->nosim != -1) {
+				if (enemy.at(popravi_nasprotnike)->nosim > (int)i && enemy.at(popravi_nasprotnike)->nosim != -1) {
 					--enemy.at(popravi_nasprotnike)->nosim;
 				}
 			}
